схлопнуть с фиксом денек

+++ src/xrGame/ui/UIActorMenu.h
@@ -211,7 +211,7 @@ protected:
 	// Controller UI
 	virtual void				SetAuxMode							(eActorMenuControllerAuxMode mode);
 
-	virtual void				TradeShowMessage					(u32 money_actor, u32 money_patner);
+	virtual void				TradeShowMessage					(s64 money_actor, s64 money_patner);
 
 public:
 								CUIActorMenu						();
